refactor(server): extract pipeStream into a StreamPipe collaborator
client-router.ts's ClientRouter mixed routing, history reconciliation,
stream piping, error persistence, artifact delivery and audit writes in
one ~2400-line class (pinchy#1073). pipeStream alone was ~680 lines,
testable only through the full WS stack.

Move pipeStream and the private helpers only it drives
(deliverRunArtifacts, persistDurableChatError, surfaceRunFailure,
writeAgentErrorAudit) into a new StreamPipe class in
packages/web/src/server/stream-pipe.ts, constructed once per
ClientRouter with the SAME shared state (activeRuns, disconnectSignal,
sessionCache, userId, userRole, openclawClient) the router already
holds, plus the router's own broadcastForRun (injected, not
duplicated, so every WS frame still goes through one readyState-gated
sender). ClientRouter's pipeStream is now a thin delegating method;
the one external writeAgentErrorAudit call site (the dispatch-race
retry hook in handleMessage) now calls streamPipe.writeAgentErrorAudit
directly.

No behavior change: event ordering, error paths, audit write patterns,
and the documented race guards (#310, #956, #583, #7, #320, #882, #703,
#978, #903, #1013) travel with their code unchanged. The move was
re-derived from main rather than merged: main landed #978 (user-abort
handling throughout the stream loop) and #903 (byte-pinned delivery
grants, plus deliverRunArtifacts' new runStartedAt parameter) INSIDE
the moved region after this branch was cut, and resolving that as a
text conflict is exactly how a refactor silently reverts two fixes.
The extraction was re-applied to main's current file by substitution
(`this.<dep>` -> injected dep) and verified by diffing the result
against the pre-rebase copy: every hunk is one of those two changes.

client-router-delivery.test.ts's private-method cast seam is updated
to reach deliverRunArtifacts through streamPipe; no assertions changed.
Two comments elsewhere pointed at the old home (pinchy-files'
deliverable-filename.ts and agent-error-classifier.ts) and now name
stream-pipe.ts.

Docs-not-needed: internal refactor, websocket behavior unchanged
