+++ packages/web/src/__tests__/server/client-router-delivery.test.ts
@@ -12,7 +12,8 @@ import { tmpdir } from "os";
 // plugin tool-output text, so this poll (not an inline marker) is how a delivery
 // is observed. The full path (real plugin → transcript artifact → this glue) is
 // covered by E2E; here we exercise the glue directly via the private-method cast
-// seam.
+// seam — routed through the StreamPipe collaborator ClientRouter delegates
+// stream-piping concerns to (pinchy#1073 step 1; see stream-pipe.ts).
 
 const { mockInsertValues, mockAppendAuditLog, mockRecordAuditFailure, mockGrantSelect } =
   vi.hoisted(() => ({
@@ -114,15 +115,17 @@ function makeRouter(artifacts: Artifact[]) {
 async function deliver(router: ClientRouter, clientWs: unknown, runStartedAt?: Date) {
   await (
     router as unknown as {
-      deliverRunArtifacts: (
-        sessionKey: string,
-        agent: { id: string; name: string },
-        clientWs: unknown,
-        messageId: string,
-        runStartedAt: Date
-      ) => Promise<void>;
+      streamPipe: {
+        deliverRunArtifacts: (
+          sessionKey: string,
+          agent: { id: string; name: string },
+          clientWs: unknown,
+          messageId: string,
+          runStartedAt: Date
+        ) => Promise<void>;
+      };
     }
-  ).deliverRunArtifacts(
+  ).streamPipe.deliverRunArtifacts(
     SESSION_KEY,
     agent,
     clientWs,
